done task

+++ pt_num.c
@@ -11,15 +11,21 @@ int ptr_num(int num)
 	int i = 0, temp = 0, count = 0;
 
 	if (num < 0)
-	{ i += _putchar('-');
-		num = -num; }
+	{ 
+		i += _putchar('-');
+		num = -num; 
+	}
 	temp = num;
 	count = 1;
 	while (temp / 10 > 0)
-	{ temp /= 10;
-		count *= 10; }
+	{ 
+		temp /= 10;
+		count *= 10; 
+	}
 	while (count > 0)
-	{ i += _putchar((num / count) % 10 + '0');
-		count /= 10; }
+	{
+		i += _putchar((unsigned int) (num / count) % 10 + '0');
+		count /= 10; 
+	}
 	return (i);
 }
